% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caRatio.R
\name{caRatio}
\alias{caRatio}
\title{caRatio}
\usage{
caRatio(m, newGraph, oldGraph, prior, scoreFun)
}
\arguments{
\item{m}{The length of the individual vector. This is the number of edges in
the graph and the fitness of the graph.}

\item{newGraph}{A vector of edge directions. This the the proposed
vector of edge directions.}

\item{oldGraph}{A vector of edge directions. This is the graph from
the start of the current interation.}

\item{prior}{A vector containing the prior probability of seeing each edge
direction.}

\item{scoreFun}{A character string indicating the fitness to be used. The
fitness can either be BIC of the log likelihood.}
}
\value{
A vector of edge directions and the fitness of the individual
}
\description{
Calculates the acceptance ratio for the Metropolis-Hastings and Genetic
Algorithm. It takes the ratio of the original vector of edge directions, old
individual, and the proposed vector of edge directions, new individual.
}
