% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carPrior.R
\name{carPrior}
\alias{carPrior}
\title{carPrior}
\usage{
carPrior(edgeDir1, edgeDir2, edgeType, pmr, prior)
}
\arguments{
\item{edgeDir1}{A scalar indicating the state of the edge for graph one.}

\item{edgeDir2}{A scalar indicating the state of the edge for graph two.}

\item{edgeType}{A 0 or 1 indicating whether the edge is a gv-ge edge (1) or
a gv-gv or ge-ge edge (1).}

\item{pmr}{Logical. If true the Metropolis-Hastings algorithm will use the
Principle of Mendelian Randomization, PMR. This prevents the direction of an
edge pointing from a gene expression node to a genetic variant node.}

\item{prior}{A vector containing the prior probability of seeing each edge
direction.}
}
\value{
The probability of the specified edge state
}
\description{
Calculates the prior depending on the current edge state and edge type for
the caRatio function. For example, gv-ge, gv-gv, ge-ge.
}
