% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cllEdge.R
\name{cllEdge}
\alias{cllEdge}
\title{cllEdge}
\usage{
cllEdge(amZero, coordinates, data, graph, m, nGV, nEdges, nNodes)
}
\arguments{
\item{amZero}{An adjacency matrix of zeros.}

\item{coordinates}{A matrix of the row and column numbers of the nonzero
elements in the adjacency matrix. The row numbers make up the first row of
the coordinates matrix and the column numbers make up the second row.}

\item{data}{A matrix with the nodes across the columns and the observations
along the rows.}

\item{graph}{The edge directions of the current graph.}

\item{m}{The length of the graph vector. This is the number of edges in the
graph plus the fitness of the graph.}

\item{nGV}{The number of genetic variants in the graph.}

\item{nEdges}{The number of edges in the graph.}

\item{nNodes}{The number of nodes in the graph.}
}
\value{
The log likelihood of the graph according to the orientation of the
edges denoted by the DNA of the current graph.
}
\description{
Calculates the log likelihood of the current node when running a genetic
algorithm on the direction of the edges in the graph.
}
