% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cllMultinom.R
\name{cllMultinom}
\alias{cllMultinom}
\title{cllMultinom}
\usage{
cllMultinom(adjMatrix, data, nGV, nNodes)
}
\arguments{
\item{adjMatrix}{The adjacency matrix according to the edge directions of the
current graph.}

\item{data}{The data matrix.}

\item{nGV}{The number of genetic variants in the graph.}

\item{nNodes}{The total number of nodes in the graph.}
}
\value{
A list containing two elements. The first element is a vector with
the log likelihood for each node. The second element is a vector with the
number of parents for each node.
}
\description{
Calculates the log likelihood for a node with multinomial data.
}
