% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coordinates}
\alias{coordinates}
\title{coordinates}
\usage{
coordinates(adjMatrix)
}
\arguments{
\item{adjMatrix}{The adjacency matrix returned by the MRPC algorithm. The
adjacency matrix is a matrix of zeros and ones. The ones represent an edge
and also indicates the direction of that edge.}
}
\value{
A matrix. The first row in the matrix holds the row locations of the
nonzero elements in the adjacency matrix and the second row holds the column
locations of the nonzero elements in the adjacency matrix.
}
\description{
Takes the adjacency matrix and extracts the row and column numbers for each
nonzero element. It also takes the nonzero elements and collapses them into a
vector. This vector is the DNA of the individual. The length of the DNA is
the number of edges in the graph.
}
