% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cPrior.R
\name{cPrior}
\alias{cPrior}
\title{cPrior}
\usage{
cPrior(edges, edgeType, pmr, prior)
}
\arguments{
\item{edges}{A vector indicating the position in the prior vector of the two
edge states that the current edge can move to.}

\item{edgeType}{A 0 or 1 indicating whether the edge is a gv-ge edge (1) or
a gv-gv or ge-ge edge (1).}

\item{pmr}{Logical. If true the Metropolis-Hastings algorithm will use the
Principle of Mendelian Randomization, PMR. This prevents the direction of an
edge pointing from a gene expression node to a genetic variant node.}

\item{prior}{A vector containing the prior probability of seeing each edge
direction.}
}
\value{
The probability of the specified edge state
}
\description{
Calculates the prior depending on the current edge state and edge type for
the mutate function. For example, gv-ge, gv-gv, ge-ge.
}
