% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleBranches.R
\name{cycleBranches}
\alias{cycleBranches}
\title{cycleBranches}
\usage{
cycleBranches(SaL)
}
\arguments{
\item{SaL}{A binary matrix with only the nodes and edges that could
potentially form a cycle.}
}
\value{
A list of matrices that contain all the cycles for each child of the
root node. The cycles are displayed across the columns of the matrices.
}
\description{
Turns the stem and leaf matrix returned by the cycleSL function into a list
of matrices. Each matrix contains all the cycles for each child of the root
node.
}
