% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleCoord.R
\name{cycleCoord}
\alias{cycleCoord}
\title{cycleCoord}
\usage{
cycleCoord(tBranches)
}
\arguments{
\item{tBranches}{A list cycles. Each element in the list is a vector of nodes
that could form a directed cycle.}
}
\value{
A list of row and column coordinates for each edge in a directed
cycle. The coordinates are to the adjacency matrix.
}
\description{
Extracts the coordinates from the adjacency matrix of each edge that is part
of a directed cycle.
}
