% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleED.R
\name{cycleED}
\alias{cycleED}
\title{cycleED}
\usage{
cycleED(tBranches)
}
\arguments{
\item{tBranches}{A list of trimmed branches. The branches in this list
contain only the nodes that could form a cycle.}
}
\value{
A list of edge directions that forms a directed cycle for each
trimmed branch passed to the cycleED function.
}
\description{
Takes the names of the nodes that could form a cycle and creates a vector of
edge directions that form a directed cycle.
}
