% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleEN.R
\name{cycleEN}
\alias{cycleEN}
\title{cycleEN}
\usage{
cycleEN(adjMatrix, cCoord)
}
\arguments{
\item{adjMatrix}{The adjacency matrix of the graph.}

\item{cCoord}{A list containing the coordinates in the adjacency matrix of
the nodes that could form a cycle}
}
\value{
A list containing the edge numbers for the edges that could form a
cycle for each cycle in the graph.
}
\description{
Gets the number of each edge in the DNA that could form a cycle.
}
