% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{cycleFndr}
\alias{cycleFndr}
\title{cycleFndr}
\usage{
cycleFndr(adjMatrix, nCPh, nGV, nEdges, pmr, position, threads = 1)
}
\arguments{
\item{adjMatrix}{The adjacency matrix of the graph.}

\item{nCPh}{The number of clinical phenotypes in the graph.}

\item{nGV}{An integer. The number of genetic variants in the graph.}

\item{nEdges}{An integer. The number of edges in the graph.}

\item{pmr}{Logical. If TRUE the principle of Mendelian randomization will be
used. This restrics any genetic variant node from being the child of a gene
expression node (when pmr = TRUE).}

\item{position}{A matrix of row and column indices for each edge in a cycle.
This is the output from the coordiantes function.}

\item{threads}{An integer specifying the number of threads to use. Default is 1}
}
\value{
A list containing the unique decimal numbers for the cycles in the
graph and the indices of each edge in every cycle. If there are no cycles
in the graph the function returns NULL.
}
\description{
Carries out all the steps to find directed cycles
}
