% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleSaL.R
\name{cycleSaL}
\alias{cycleSaL}
\title{cycleSaL}
\usage{
cycleSaL(adjMatrix)
}
\arguments{
\item{adjMatrix}{The adjacency matrix of the graph.}
}
\value{
A binary matrix where the 1s represent an edge between the row node
and the column node. If there are no cycles in the graph the function returns
NULL.
}
\description{
The first step in identifying potential directed cycles in the graph is to
find the nodes that could form a cycle. Since a node must be connected to at
least two other nodes to be a part of a cycle we start by summing all of the
rows of the undirected adjacency matrix and creating a new matrix that only
has the rows of the undirected adjacency matrix with a sum of two or higher.
The function cycleSaL creates the matrix of nodes that could form a directed
cycle.
}
\details{
Creates a stem and leaf matrix from the adjacency matrix with the nodes that
could potentially form a cycle. The row names of the matrix created by the
cycleSaL function are the stems and the 1s in the matrix are the leaves.
}
