% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal.R
\name{decimal}
\alias{decimal}
\title{decimal}
\usage{
decimal(edgeDir, edgeNum)
}
\arguments{
\item{edgeDir}{A list containing the directions of the edges for each cycle
in the graph.}

\item{edgeNum}{A list containing the edge index of each edge in each cycle
of the graph.}
}
\value{
A vector containing the decimal number of each cycle.
}
\description{
Turns the directions and locations of the cycles in the graph in to a decimal
number.
}
