% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimalCI.R
\name{decimalCI}
\alias{decimalCI}
\title{decimalCI}
\usage{
decimalCI(edgeNum, individual)
}
\arguments{
\item{edgeNum}{A list containing the index of each edge for each cycle in
the graph.}

\item{individual}{A vector containing the DNA and the log likelihood of the
current individual.}
}
\value{
The decimal numbers for each potential cycle given the edge
directions of the current individual.
}
\description{
Calculates the decimal number of the cycles of an individual according to its
current edge directions.
}
