% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detType.R
\name{detType}
\alias{detType}
\title{detType}
\usage{
detType(coord, nEdges, nGV, pmr)
}
\arguments{
\item{coord}{A matrix of row and column indices for each edge in the
network. The indices are from the adjacency matrix.}

\item{nEdges}{The number of edges in the graph.}

\item{nGV}{The number of genetic variants in the graph.}

\item{pmr}{Logical. If true the Metropolis-Hastings algorithm will use the
Principle of Mendelian Randomization, PMR. This prevents the direction of an
edge pointing from a gene expression node to a genetic variant node.}
}
\value{
A vector of edge types: 1 if edge is between gv-ge and  0 otherwise.
}
\description{
Determines the type of edge between two nodes. For example, gv-ge, gv-gv,
ge-ge.
}
