% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efGA.R
\name{efGA}
\alias{efGA}
\title{efGA}
\usage{
efGA(evolvedPop)
}
\arguments{
\item{evolvedPop}{A list of evolved populations.}
}
\value{
A matrix. The number of rows is equal to the number of edges in the
graph. It contains two columns. The first column is the frequency of the edge
whose directions goes from the node with a smaller index to the node with a
larger index. The second column contains the frequencies of the edges whose
direction is opposite from the first column.
}
\description{
Calculates the number of times each edge is oriented in each of the two
possible directions.
}
