% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efMH.R
\name{efMH}
\alias{efMH}
\title{efMH}
\usage{
efMH(dEdge)
}
\arguments{
\item{dEdge}{A matrix from the output of the MHEdges function.}
}
\value{
A matrix. The number of rows is equal to the number of edges in the
graph. It contains three columns. The first column is the frequency of the
edge whose directions goes from the node with a smaller index to the node
with a larger index. The second column contains the frequencies of the edges
whose direction is opposite from the first column. The third column is the
frequency of not having that edge in the graph.
}
\description{
Calculates the number of times each edge is oriented in each of the two
possible directions.
}
