% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates}
\alias{estimates}
\title{estimates}
\usage{
estimates(adjMatrix, node, data)
}
\arguments{
\item{adjMatrix}{The adjacency matrix created from the DNA of the current
individual.}

\item{node}{The column number of the current node, the node for which we will
calculate the estimates for the mean and standard deviation.}

\item{data}{A matrix with the nodes across the columns and the observations
along the rows.}
}
\value{
A list containing the number of parents of the current node, a vector
with the estimates of the beta coefficients and standard error of the
residuals, and the data matrix containing only the columns of the current
node and its parents.
}
\description{
Calculates the estimates of the mean and standard deviation for the current
node. If the current node doesn't have any parent then the mean and standard
deviation are simply calculated from the data. If the current node has at
least one parent the mean is estimated by a linear model with the current
node as the independent variable and parent nodes as independent variables.
The standard deviation is estimated by the standard error of the residuals.
These estimates are later used to calculate the log likelihood of the current
node.
}
