% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatesM.R
\name{estimatesM}
\alias{estimatesM}
\title{estimatesM}
\usage{
estimatesM(adjMatrix, data, node)
}
\arguments{
\item{adjMatrix}{The adjacency matrix created from the edge directions of the
current graph.}

\item{data}{A matrix with the nodes across the columns and the observations
down the rows.}

\item{node}{The column number of the current node, the node for which we will
calculate the estimates for the mean and standard deviation.}
}
\value{
A list containing the number of parents of the current node, a vector
with the estimates of the beta coefficients, and the data matrix containing
only the columns of the current node and its parents.
}
\description{
Calculates the estimates of the probability of being in each group of the
current genetic variant.
}
