% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolveEdges.R
\name{evolveEdges}
\alias{evolveEdges}
\title{evolveEdges}
\usage{
evolveEdges(N, adjMatrix, data, tournamentSize, mutationRate, generations,
  scoreFun)
}
\arguments{
\item{N}{An integer for the number of individuals to generate.}

\item{adjMatrix}{The adjacency matrix returned by the MRPC algorithm. The
adjacency matrix is a matrix of zeros and ones. The ones represent an edge
and also indicates the direction of that edge.}

\item{data}{A matrix with the nodes across the columns and the observations
along the rows.}

\item{tournamentSize}{An integer. The number of individuals that will be
passed to the selectBest and selectWorst functions.}

\item{mutationRate}{A number between 0 and 1. As a rule of thumb the mutation
rate should be 1 / length of the DNA.}

\item{generations}{The number of times the individual with the highest log
likelihood is selected, cloned, and mutated.}

\item{scoreFun}{A character string indicating what method to use for
calculating the fitness.}
}
\value{
A matrix where the first m columns are the DNA and the m + 1 column
is the log likelihood. Each row represents one individual.
}
\description{
evolveEdges
}
