% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topologies.R
\name{gmac_s4}
\alias{gmac_s4}
\title{gmac_s4}
\usage{
gmac_s4(N, p, beta_h, beta_lc, beta_lt)
}
\arguments{
\item{N}{The number of observations to generate.}

\item{p}{The frequency of the reference allele.}

\item{beta_h}{The signal strength of the hidden variables on the cis- and
trans- genes.}

\item{beta_lc}{The signal strength of the eQTL on the cis-gene.}

\item{beta_lt}{The signal strength of the eQTL on the trans-gene.}
}
\description{
Simulates data under scenario four with two hidden variables from the gmac
algorithm paper.
}
