% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topologies.R
\name{m1_hidden_c}
\alias{m1_hidden_c}
\title{m1_hidden_c}
\usage{
m1_hidden_c(N, p, beta_h, beta_t1, beta_t2)
}
\arguments{
\item{N}{The number of observations to generate.}

\item{p}{The frequency of the reference allele.}

\item{beta_h}{The signal strength of the hidden variables on T1 and T2.}

\item{beta_t1}{The signal strength of V on T1.}

\item{beta_t2}{The signal strength of T1 on T2.}
}
\description{
Simulates data under topology M1 with two hidden (confounder) variables.
}
