% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topologies.R
\name{m2gv}
\alias{m2gv}
\title{m2gv}
\usage{
m2gv(N, p, ss)
}
\arguments{
\item{N}{The number of observations to generate.}

\item{p}{The frequency of the reference allele.}

\item{ss}{The value of the beta_1 coefficient in the linear model
beta_0 + beta_1 * x_1. This term is referred to as the signal strength.}
}
\description{
Simulates data under topology M2 with one genetic variant node.
}
