% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFormula.R
\name{makeFormula}
\alias{makeFormula}
\title{makeFormula}
\usage{
makeFormula(estimates, nParents, parentData)
}
\arguments{
\item{estimates}{The estimates of the linear model calculated in the
estimates function.}

\item{nParents}{The number of parents of the current node}

\item{parentData}{The data matrix containing only the columns of the
current node and the parents of the current node.}
}
\value{
A vector of means calculated by using the estimates from the
estimates function.
}
\description{
Takes the estimates calculated in the linear model from the estimates
function and creates a linear model to pass as the mean to the dnorm
function.
}
