% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{mcmc-class}
\alias{mcmc-class}
\title{mcmc class}
\description{
mcmc class
}
\section{Slots}{

\describe{
\item{\code{chain}}{A matrix where the rows contain the vector of edge states for the
graph that is accepted at each iteration.}

\item{\code{decimal}}{A vector where each element is the decimal number of the graph
accepted at each iteration of the MH algorithm.}

\item{\code{likelihood}}{A vector where each element is the likelihood of the graph
accepted at each iteration of the MH algorithm.}

\item{\code{time}}{The time in seconds it took the MH algorithm to run.}
}}

