% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topologies.R
\name{mpge}
\alias{mpge}
\title{mpge}
\usage{
mpge(N, ss)
}
\arguments{
\item{N}{The number of observations to generate.}

\item{ss}{The value of the beta_1 coefficient in the linear model
beta_0 + beta_1 * x_1. This term is referred to as the signal strength.}
}
\description{
Simulates data under the multi-parent topology.
}
