% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{multiplot}
\usage{
multiplot(..., cols = 1)
}
\arguments{
\item{...}{The graphs to be plotted on one page.}

\item{cols}{The number of columns on the plot page.}
}
\description{
A function from the Cookbook for R website that prints multiple plots on one
page.
}
