% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\title{mutate}
\usage{
mutate(graph, m, mutationRate, prior)
}
\arguments{
\item{graph}{A vector containing the DNA of the graph who was
selected by the selectBest function.}

\item{m}{The length of the graph vector. This is the number of edges in the
graph plus the fitness of the graph.}

\item{mutationRate}{A number between 0 and 1. As a rule of thumb the mutation
rate should be 1 / length of the DNA.}

\item{prior}{A vector containing the prior probability of seeing each edge
direction.}
}
\value{
A vector of the edge directions and log likelihood.
}
\description{
Takes the edge directions of the current graph mutates it according to the
mutation rate passed to the function.
}
