% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot,baycn,ANY-method}
\alias{plot,baycn,ANY-method}
\alias{plot,baycn-method}
\title{plot}
\usage{
\S4method{plot}{baycn,ANY}(x, y, ...)
}
\arguments{
\item{x}{An object of classs baycn.}

\item{y}{Optional if x is the appropriate structure.}

\item{...}{Other Arguments passed to methods.}
}
\description{
plot
}
