% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plot,mcmc,ANY-method}
\alias{plot,mcmc,ANY-method}
\alias{plot,mcmc-method}
\title{plot}
\usage{
\S4method{plot}{mcmc,ANY}(x, y, ...)
}
\arguments{
\item{x}{An object of classs mcmc.}

\item{y}{Optional if x is the appropriate structure.}

\item{...}{Other Arguments passed to methods.}
}
\description{
plot
}
