% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate.R
\name{populate}
\alias{populate}
\title{populate}
\usage{
populate(N, adjMatrix, data, dnUnique, edgeNum, wCycle, scoreFun)
}
\arguments{
\item{N}{An integer for the number of individuals to generate.}

\item{adjMatrix}{The adjacency matrix returned by the MRPC algorithm. The
adjacency matrix is a matrix of zeros and ones. The ones represent an edge
and also indicates the direction of that edge.}

\item{data}{A matrix with the nodes across the columns and the observations
along the rows.}

\item{dnUnique}{A vector of unique decimal numbers. These numbers represent
each cycle in the graph. This argument is NULL if there are no cycles in the
graph.}

\item{edgeNum}{A list containing the number of each edge in each of the
cycles in the graph. This argument is NULL if there are no cycles in the
graph.}

\item{wCycle}{A vector of the locations in the cycleDN vector where each
unique number occurs. The unique numbers refer to each of the unique cycles
in the graph. This argument is NULL if there are no cycles in the graph.}

\item{scoreFun}{A character string indicating what method to use for
calculating the fitness.}
}
\value{
A list where the first element of the list is the population matrix
and the second element is the coordinates matrix created by the
coordinates function.
}
\description{
Creates a matrix where the first m columns are the DNA and the m + 1 column
is the log likelihood. Each row represents one individual.
}
