% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prerec.R
\name{prerec}
\alias{prerec}
\title{prerec}
\usage{
prerec(iMatrix, tMatrix)
}
\arguments{
\item{iMatrix}{The undirected adjacency matrix of the inferred graph.}

\item{tMatrix}{The undirected adjacency matrix of the true graph.}
}
\value{
A list. The first element is the precision and the second element is
the recall of the inferred graph.
}
\description{
Calculates the precision and recall of the inferred graph.
}
