% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{rMParents}
\alias{rMParents}
\title{rMParents}
\usage{
rMParents(N, mParents, parentData, b0, b1, s)
}
\arguments{
\item{N}{The number of observations to simulate.}

\item{mParents}{The number of parents of the current gene.}

\item{parentData}{A list containing the data of the parents of the current
gene.}

\item{b0}{The slope of the linear model
b0 + b1[[1]] * parentData[[1]] + b1[[2]] * parentData[[2]] + ...}

\item{b1}{A list containing the regression coefficients of the parents of the
gene for which data is being simulated.}

\item{s}{The standard deviation of the gene for which data is being
simulated.}
}
\description{
rMParents
}
