% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{rNoParents}
\alias{rNoParents}
\title{rNoParents}
\usage{
rNoParents(N, b0, s)
}
\arguments{
\item{N}{The number of observations to simulate.}

\item{b0}{The mean of the normal distribution.}

\item{s}{The standard deviation of the normal distribution.}
}
\value{
A vector containing the simulated data.
}
\description{
Creates data from a normal distribution. This function is used for generating
data for nodes with no parents.
}
