% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.R
\name{selectBest}
\alias{selectBest}
\title{selectBest}
\usage{
selectBest(population, tournamentSize, scoreFun)
}
\arguments{
\item{population}{A matrix where the first m columns are the DNA of the
individuals. Where m is the number of edges in the graph. The final column
is the log likelihood of the individual. The number of rows is the population
size.}

\item{tournamentSize}{An integer. This is the number of individuals that will
be selected from the population and a winner, the individual with the highest
log likelihood, will be selected to be cloned and mutated. Then it will
replace the individual selected by the selectWorst function. In general the
tournament size is around one tenth of the population size.}

\item{scoreFun}{A character string indicating what method to use for
calculating the fitness.}
}
\value{
A vector of length m. Where m is the number of edges in the graph.
}
\description{
Selects the individual with the highest log likelihood among a small subset
of the entire population.
}
