% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.R
\name{selectWorst}
\alias{selectWorst}
\title{selectWorst}
\usage{
selectWorst(population, tournamentSize, scoreFun)
}
\arguments{
\item{population}{A matrix where the first m columns are the DNA of the
individuals. Where m is the number of edges in the graph. The final column
is the log likelihood of the individual. The number of rows is the population
size.}

\item{tournamentSize}{An integer. This is the number of individuals that will
be selected from the population and a loser, the individual with the lowest
log likelihood, will be replaced by the individual that was selected by the
selectBest function after the winner was cloned and mutated. In general the
tournament size is around one tenth of the population size.}

\item{scoreFun}{A character string indicating what method to use for
calculating the fitness.}
}
\value{
A vector of length m. Where m is the number of edges in the graph.
}
\description{
Selects the individual with the lowest log likelihood among a small subset
of the entire population.
}
