% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary,baycn-method}
\alias{summary,baycn-method}
\title{summary}
\usage{
\S4method{summary}{baycn}(object, ...)
}
\arguments{
\item{object}{An object of class baycn.}

\item{...}{Other Arguments passed to methods.}
}
\description{
summary
}
