% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{summary,mcmc-method}
\alias{summary,mcmc-method}
\title{summary}
\usage{
\S4method{summary}{mcmc}(object, ...)
}
\arguments{
\item{object}{An object of class mcmc.}

\item{...}{Other Arguments passed to methods.}
}
\description{
summary
}
