% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toDecimal.R
\name{toDecimal}
\alias{toDecimal}
\title{toDecimal}
\usage{
toDecimal(edgeDir)
}
\arguments{
\item{edgeDir}{A matrix of the posterior distribution sample.}
}
\value{
A vector containing the decimal number of each graph.
}
\description{
Turns the edge directions of the graph into a decimal number.
}
