% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toED.R
\name{toED}
\alias{toED}
\title{toED}
\usage{
toED(adjMatrix, coordinates, graph, nGV, nNodes)
}
\arguments{
\item{adjMatrix}{The adjacency matrix returned by the MRPC algorithm. The
adjacency matrix is a matrix of zeros and ones. The ones represent an edge
and also indicates the direction of that edge.}

\item{coordinates}{A matrix of the row and column numbers of the nonzero
elements in the adjacency matrix. The row numbers make up the first row of
the coordinates matrix and the column numbers make up the second row.}

\item{graph}{A vector containing the edge directions of the current graph.}

\item{nGV}{The number of genetic variants in the graph.}

\item{nNodes}{The number of nodes in the graph.}
}
\value{
A vector of 0s, 1s, and 2s, indicating the directions of the edges
for the current graph.
}
\description{
Takes the adjacency matrix when there are genetic variants in the data and
removes the edges if there are any genetic variant nodes that have parents
that are gene expression nodes.
}
