% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracePlot.R
\name{tracePlot}
\alias{tracePlot}
\title{tracePlot}
\usage{
tracePlot(x)
}
\arguments{
\item{x}{An object of class baycn.}
}
\description{
Creates a trace plot of the log likelihood and graph decimal. The graph
decimal is a decimal number calculated from the vector of edge states for the
accepted graph.
}
