% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caRatio.R
\name{caRatio}
\alias{caRatio}
\title{caRatio}
\usage{
caRatio(current, edgeType, m, pmr, proposed, prior)
}
\arguments{
\item{current}{A vector of edge directions. This is the graph from
the start of the current interation.}

\item{edgeType}{A 0 or 1 indicating whether the edge is a gv-ge edge (0) or
a gv-gv or ge-ge edge (1).}

\item{m}{The length of the individual vector. This is the number of edges in
the graph and the fitness of the graph.}

\item{pmr}{Logical. If true the Metropolis-Hastings algorithm will use the
Principle of Mendelian Randomization, PMR. This prevents the direction of an
edge pointing from a gene expression node to a genetic variant node.}

\item{proposed}{A vector of edge directions. This the the proposed
vector of edge directions.}

\item{prior}{A vector containing the prior probability of seeing each edge
direction.}
}
\value{
A vector of edge directions and the fitness of the individual
}
\description{
Calculates the acceptance ratio for the Metropolis-Hastings and Genetic
Algorithm. It takes the ratio of the original vector of edge directions, old
individual, and the proposed vector of edge directions, new individual.
}
