% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cllEdge.R
\name{cllEdge}
\alias{cllEdge}
\title{cllEdge}
\usage{
cllEdge(individual, coordinates, data, scoreFun)
}
\arguments{
\item{individual}{The DNA of the current individual.}

\item{coordinates}{A matrix of the row and column numbers of the nonzero
elements in the adjacency matrix. The row numbers make up the first row of
the coordinates matrix and the column numbers make up the second row.}

\item{data}{A matrix with the nodes across the columns and the observations
along the rows.}

\item{scoreFun}{A character string indicating what method to use for
calculating the fitness.}
}
\value{
The log likelihood of the graph according to the orientation of the
edges denoted by the DNA of the current individual.
}
\description{
Calculates the log likelihood of the current node when running a genetic
algorithm on the direction of the edges in the graph.
}
