% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cllNormal.R
\name{cllNormal}
\alias{cllNormal}
\title{cllNormal}
\usage{
cllNormal(adjMatrix, data, logLikelihood, nGV, nNodes)
}
\arguments{
\item{adjMatrix}{The adjacency matrix according to the edge directions of the
current graph.}

\item{data}{The data matrix.}

\item{logLikelihood}{A vector containing the log likelihood for each genetic
variant node.}

\item{nGV}{The number of genetic variants in the graph.}

\item{nNodes}{The total number of nodes in the graph.}
}
\value{
A vector with the log likelihood for each node.
}
\description{
Calculates the log likelihood for a node with normal data.
}
