% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleBranches.R
\name{cycleBranches}
\alias{cycleBranches}
\title{cycleBranches}
\usage{
cycleBranches(SaL)
}
\arguments{
\item{SaL}{A binary matrix containing the nodes that could potentially form a
directed cycle.}
}
\value{
A matrix that contains all the cycles for the current graph. The
cycles are displayed down the columns of the matrix.
}
\description{
Turns the stem and leaf matrix returned by the cycleSaL function into a list
of matrices. Each matrix contains all the cycles for each child of the root
node.
}
