% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleCoord.R
\name{cycleCoord}
\alias{cycleCoord}
\title{cycleCoord}
\usage{
cycleCoord(tBranches)
}
\arguments{
\item{tBranches}{A list of the nodes that could form a cycle.}
}
\value{
A list of row column coordinates for each node that could form a
cycle. The coordinates are to the adjacency matrix.
}
\description{
Gets the coordinates in the adjacency matrix of the nodes that could form a
cycle.
}
