% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclePrep.R
\name{cyclePrep}
\alias{cyclePrep}
\title{cyclePrep}
\usage{
cyclePrep(adjMatrix)
}
\arguments{
\item{adjMatrix}{The adjacency matrix of the graph.}
}
\value{
A list containing the unique decimal numbers for the cycles in the
graph, the number of each edge in each cycle, and the location in the
original cycleDN list where each unique cycle occurs. If there are no cycles
in the graph the function returns NULL.
}
\description{
Carries out all of the steps needed in order to remove a cycle.
}
