% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleTB.R
\name{cycleTB}
\alias{cycleTB}
\title{cylceTB}
\usage{
cycleTB(branches)
}
\arguments{
\item{branches}{A matrix that holds all the branches of a tree.}
}
\value{
A list. The first element, tBranches, is a list of trimmed
branches. The second element, nBranches, is an integer of the total number of
branches in the tree.
}
\description{
Takes the branches created by the cycleBranches function and trims them. The
cycleBranches function creates a tree as deep as possible. Therefore, some of
the branches may have nodes that are not part of a cycle.
}
