% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\title{mutate}
\usage{
mutate(edgeType, graph, nEdges, pmr, prior, ztbProb)
}
\arguments{
\item{edgeType}{A 0 or 1 indicating whether the edge is a gv-ge edge (1) or
a gv-gv or ge-ge edge (0).}

\item{graph}{A vector containing the current edge states for each edge.}

\item{nEdges}{The number of edges in the graph.}

\item{pmr}{Logical. If true the Metropolis-Hastings algorithm will use the
Principle of Mendelian Randomization, PMR. This prevents the direction of an
edge pointing from a gene expression node to a genetic variant node.}

\item{prior}{A vector containing the prior probability of seeing each edge
direction.}

\item{ztbProb}{A vector of probabilities for 1:nEdges from a zero truncated
binomial distribution.}
}
\value{
A vector of the edge directions and log likelihood.
}
\description{
Takes the current graph and changes at least one of the edges from its
current edge state to a different edge state.
}
